Move cache methods into new file
